+++ lib/easytest/case.rb
@@ -9,9 +9,9 @@ module Easytest
     alias skipped? skipped
     alias only? only
 
-    def initialize(name:, file:, skipped: false, only: false, &block)
+    def initialize(name:, skipped: false, only: false, &block)
       @name = name
-      @file = file
+      @file = caller_locations(3, 1).first.absolute_path
       @block = block
       @skipped = skipped
       @only = only