+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add `before` and `after` hooks.
 - Add `to_have_attributes` matcher.
 - Add `to_satisfy` matcher.